Split out graph from svg creation
